+++ src/main/java/com/roome/roome/be/domain/user/dto/social/GoogleTokenResponse.java
@@ -1,15 +1,12 @@
-package com.roome.roome.be.domain.user.dto.kakao;
+package com.roome.roome.be.domain.user.dto.social;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-// 카카오 토큰 응답
 @Getter
 @NoArgsConstructor
-public class KakaoTokenResponse {
-
-    @JsonProperty("token_type")
-    private String tokenType;
+public class GoogleTokenResponse {
 
     @JsonProperty("access_token")
     private String accessToken;
@@ -20,7 +17,12 @@ public class KakaoTokenResponse {
     @JsonProperty("refresh_token")
     private String refreshToken;
 
-    @JsonProperty("refresh_token_expires_in")
-    private Integer refreshTokenExpiresIn;
-}
+    @JsonProperty("scope")
+    private String scope;
 
+    @JsonProperty("token_type")
+    private String tokenType;
+
+    @JsonProperty("id_token")
+    private String idToken;
+}
