+++ src/main/java/com/roome/roome/be/domain/user/dto/social/KakaoTokenResponse.java
@@ -1,12 +1,15 @@
-package com.roome.roome.be.domain.user.dto.google;
-
+package com.roome.roome.be.domain.user.dto.social;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+// 카카오 토큰 응답
 @Getter
 @NoArgsConstructor
-public class GoogleTokenResponse {
+public class KakaoTokenResponse {
+
+    @JsonProperty("token_type")
+    private String tokenType;
 
     @JsonProperty("access_token")
     private String accessToken;
@@ -17,12 +20,7 @@ public class GoogleTokenResponse {
     @JsonProperty("refresh_token")
     private String refreshToken;
 
-    @JsonProperty("scope")
-    private String scope;
-
-    @JsonProperty("token_type")
-    private String tokenType;
-
-    @JsonProperty("id_token")
-    private String idToken;  
+    @JsonProperty("refresh_token_expires_in")
+    private Integer refreshTokenExpiresIn;
 }
+
